requirement.txt
selenium
pytest
allure-pytest
pytest-xdist
pytest-rerunfailures
pytest-ordering
pillow
pywinauto
ddddocr
pyyaml
requests
opencv-python